+++ app-hooks/post-installation/00-setup-trusted-domains.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+set -euox pipefail
+
+php /var/www/html/occ config:system:set trusted_domains 2 --value=host.docker.internal